use_cases/finetuning_selection/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.24.0

# Feature extraction
sentence-transformers>=2.2.0
spacy>=3.5.0

# Evaluation
lm-eval>=0.4.0

# Visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0

# For local LLM judge (optional)
# vllm>=0.2.0  # For faster inference if needed
